+++ packages/wallet/src/initialization/instances/remote-feature-flag-controller/remote-feature-flag-controller.ts
@@ -17,6 +17,7 @@ export const remoteFeatureFlagController: InitializationConfiguration<
       messenger,
       clientConfigApiService: options.clientConfigApiService,
       getMetaMetricsId: options.getMetaMetricsId ?? ((): string => ''),
+      getCanonicalId: options.getCanonicalId ?? ((): string => ''),
       clientVersion: options.clientVersion ?? '0.0.0',
       prevClientVersion: options.prevClientVersion,
       fetchInterval: options.fetchInterval,